+++ projects/packages/forms/src/dashboard/inbox/single-actions-menu.js
@@ -44,7 +44,7 @@ const SingleActionsMenu = ( { id } ) => {
 				{ append: true }
 			);
 			setLoading( false );
-		} catch ( error ) {
+		} catch {
 			setLoading( false );
 		}
 	};