+++ projects/packages/my-jetpack/_inc/components/product-card/status.tsx
@@ -19,6 +19,10 @@ const getStatusLabel: StatusStateFunction = ( status, isOwned ) => {
 		case PRODUCT_STATUSES.ACTIVE:
 		case PRODUCT_STATUSES.CAN_UPGRADE:
 			return __( 'Active', 'jetpack-my-jetpack' );
+		case PRODUCT_STATUSES.EXPIRING_SOON:
+			return __( 'Expires soon', 'jetpack-my-jetpack' );
+		case PRODUCT_STATUSES.EXPIRED:
+			return __( 'Expired plan', 'jetpack-my-jetpack' );
 		case PRODUCT_STATUSES.INACTIVE:
 		case PRODUCT_STATUSES.MODULE_DISABLED:
 		case PRODUCT_STATUSES.NEEDS_ACTIVATION:
@@ -36,6 +40,8 @@ const getStatusLabel: StatusStateFunction = ( status, isOwned ) => {
 			const inactiveText = __( 'Inactive', 'jetpack-my-jetpack' );
 			return isOwned ? needsPlanText : inactiveText;
 		}
+		case PRODUCT_STATUSES.NEEDS_ATTENTION:
+			return __( 'Needs attention', 'jetpack-my-jetpack' );
 		default:
 			return __( 'Inactive', 'jetpack-my-jetpack' );
 	}
@@ -49,6 +55,7 @@ const getStatusClassName: StatusStateFunction = ( status, isOwned ) => {
 		case PRODUCT_STATUSES.ABSENT_WITH_PLAN:
 		case PRODUCT_STATUSES.SITE_CONNECTION_ERROR:
 		case PRODUCT_STATUSES.USER_CONNECTION_ERROR:
+		case PRODUCT_STATUSES.EXPIRING_SOON:
 			return styles.warning;
 		case PRODUCT_STATUSES.INACTIVE:
 		case PRODUCT_STATUSES.NEEDS_FIRST_SITE_CONNECTION:
@@ -56,6 +63,9 @@ const getStatusClassName: StatusStateFunction = ( status, isOwned ) => {
 			return styles.inactive;
 		case PRODUCT_STATUSES.NEEDS_PLAN:
 			return isOwned ? styles.warning : styles.inactive;
+		case PRODUCT_STATUSES.EXPIRED:
+		case PRODUCT_STATUSES.NEEDS_ATTENTION:
+			return styles.error;
 		default:
 			return styles.inactive;
 	}