+++ projects/packages/my-jetpack/global.d.ts
@@ -22,7 +22,10 @@ type ProductStatus =
 	| 'needs_activation'
 	| 'needs_first_site_connection'
 	| 'user_connection_error'
-	| 'can_upgrade';
+	| 'can_upgrade'
+	| 'needs_attention'
+	| 'expired'
+	| 'expiring';
 
 type JetpackModule =
 	| 'anti-spam'
@@ -40,7 +43,8 @@ type JetpackModule =
 	| 'protect'
 	| 'videopress'
 	| 'stats'
-	| 'growth';
+	| 'growth'
+	| 'complete';
 
 type ThreatItem = {
 	// Protect API properties (free plan)
@@ -68,6 +72,29 @@ type ScanItem = {
 	version: string;
 };
 
+type RewindStatus =
+	| 'missing_plan'
+	| 'no_connected_jetpack'
+	| 'no_connected_jetpack_with_credentials'
+	| 'vp_active_on_site'
+	| 'vp_can_transfer'
+	| 'host_not_supported'
+	| 'multisite_not_supported'
+	| 'no_site_found';
+
+type BackupStatus =
+	| 'started'
+	| 'finished'
+	| 'no-credentials'
+	| 'backups-deactivated'
+	| 'no-credentials-atomic'
+	| 'credential-error'
+	| 'http-only-error'
+	| 'not-accessible'
+	| 'backup-deactivated'
+	| 'Kill switch active'
+	| 'error'
+	| 'error-will-retry';
 interface Window {
 	myJetpackInitialState?: {
 		siteSuffix: string;
@@ -156,6 +183,8 @@ interface Window {
 					plugin_slug: string;
 					post_activation_url: string;
 					post_checkout_url?: string;
+					manage_paid_plan_purchase_url?: string;
+					renew_paid_plan_purchase_url?: string;
 					pricing_for_ui?: {
 						available: boolean;
 						wpcom_product_slug: string;
@@ -173,6 +202,11 @@ interface Window {
 							should_prorate_when_offer_ends: boolean;
 							transition_after_renewal_count: number;
 							usage_limit?: number;
+							reason?: {
+								errors: {
+									introductoryOfferRemovedSubscriptionFound: string[];
+								};
+							};
 						};
 						tiers?: {
 							[ key: string ]: {
@@ -187,6 +221,11 @@ interface Window {
 									shouldProrateWhenOfferEnds: boolean;
 									transitionAfterRenewalCount: number;
 									usageLimit?: number;
+									reason?: {
+										errors: {
+											introductoryOfferRemovedSubscriptionFound: string[];
+										};
+									};
 								};
 								isIntroductoryOffer: boolean;
 								productTerm: string;
@@ -347,6 +386,27 @@ interface Window {
 				data: {
 					plugin: string;
 				};
+			};
+			backup_failure?: {
+				source: 'rewind' | 'last_backup';
+				status: RewindStatus | BackupStatus;
+				last_updated: string;
+			};
+			[ key: `${ string }--plan_expired` ]: {
+				product_slug: string;
+				product_name?: string;
+				expiry_date?: string;
+				expiry_message?: string;
+				manage_url?: string;
+				products_effected?: string[];
+			};
+			[ key: `${ string }--plan_expiring_soon` ]: {
+				product_slug: string;
+				product_name?: string;
+				expiry_date?: string;
+				expiry_message?: string;
+				manage_url?: string;
+				products_effected?: string[];
 			};
 		};
 		recommendedModules: {