+++ projects/packages/my-jetpack/src/class-products.php
@@ -20,6 +20,8 @@ class Products {
 	const STATUS_USER_CONNECTION_ERROR       = 'user_connection_error';
 	const STATUS_ACTIVE                      = 'active';
 	const STATUS_CAN_UPGRADE                 = 'can_upgrade';
+	const STATUS_EXPIRING_SOON               = 'expiring';
+	const STATUS_EXPIRED                     = 'expired';
 	const STATUS_INACTIVE                    = 'inactive';
 	const STATUS_MODULE_DISABLED             = 'module_disabled';
 	const STATUS_PLUGIN_ABSENT               = 'plugin_absent';
@@ -27,6 +29,7 @@ class Products {
 	const STATUS_NEEDS_PLAN                  = 'needs_plan';
 	const STATUS_NEEDS_ACTIVATION            = 'needs_activation';
 	const STATUS_NEEDS_FIRST_SITE_CONNECTION = 'needs_first_site_connection';
+	const STATUS_NEEDS_ATTENTION             = 'needs_attention';
 
 	/**
 	 * List of statuses that display the module as disabled
@@ -64,6 +67,7 @@ class Products {
 		self::STATUS_USER_CONNECTION_ERROR,
 		self::STATUS_PLUGIN_ABSENT_WITH_PLAN,
 		self::STATUS_NEEDS_PLAN,
+		self::STATUS_NEEDS_ATTENTION,
 	);
 
 	/**
@@ -77,6 +81,16 @@ class Products {
 	);
 
 	/**
+	 * List of statuses that display the module as active
+	 *
+	 * @var array
+	 */
+	public static $expiring_or_expired_module_statuses = array(
+		self::STATUS_EXPIRING_SOON,
+		self::STATUS_EXPIRED,
+	);
+
+	/**
 	 * List of all statuses that a product can have
 	 *
 	 * @var array
@@ -86,6 +100,8 @@ class Products {
 		self::STATUS_USER_CONNECTION_ERROR,
 		self::STATUS_ACTIVE,
 		self::STATUS_CAN_UPGRADE,
+		self::STATUS_EXPIRING_SOON,
+		self::STATUS_EXPIRED,
 		self::STATUS_INACTIVE,
 		self::STATUS_MODULE_DISABLED,
 		self::STATUS_PLUGIN_ABSENT,
@@ -93,6 +109,7 @@ class Products {
 		self::STATUS_NEEDS_PLAN,
 		self::STATUS_NEEDS_ACTIVATION,
 		self::STATUS_NEEDS_FIRST_SITE_CONNECTION,
+		self::STATUS_NEEDS_ATTENTION,
 	);
 
 	/**
@@ -105,22 +122,28 @@ class Products {
 	 */
 	public static function get_products_classes() {
 		$classes = array(
-			'anti-spam'  => Products\Anti_Spam::class,
-			'backup'     => Products\Backup::class,
-			'boost'      => Products\Boost::class,
-			'crm'        => Products\Crm::class,
-			'creator'    => Products\Creator::class,
-			'extras'     => Products\Extras::class,
-			'jetpack-ai' => Products\Jetpack_Ai::class,
-			'scan'       => Products\Scan::class,
-			'search'     => Products\Search::class,
-			'social'     => Products\Social::class,
-			'security'   => Products\Security::class,
-			'protect'    => Products\Protect::class,
-			'videopress' => Products\Videopress::class,
-			'stats'      => Products\Stats::class,
-			'ai'         => Products\Jetpack_Ai::class,
-			'growth'     => Products\Growth::class,
+			'anti-spam'        => Products\Anti_Spam::class,
+			'backup'           => Products\Backup::class,
+			'boost'            => Products\Boost::class,
+			'crm'              => Products\Crm::class,
+			'creator'          => Products\Creator::class,
+			'extras'           => Products\Extras::class,
+			'jetpack-ai'       => Products\Jetpack_Ai::class,
+			// TODO: Remove this duplicate class ('ai')? See: https://github.com/Automattic/jetpack/pull/35910#pullrequestreview-2456462227
+			'ai'               => Products\Jetpack_Ai::class,
+			'scan'             => Products\Scan::class,
+			'search'           => Products\Search::class,
+			'social'           => Products\Social::class,
+			'security'         => Products\Security::class,
+			'protect'          => Products\Protect::class,
+			'videopress'       => Products\Videopress::class,
+			'stats'            => Products\Stats::class,
+			'growth'           => Products\Growth::class,
+			'complete'         => Products\Complete::class,
+			// Features
+			'newsletter'       => Products\Newsletter::class,
+			'site-accelerator' => Products\Site_Accelerator::class,
+			'related-posts'    => Products\Related_Posts::class,
 		);
 
 		/**
@@ -342,7 +365,7 @@ class Products {
 			'protect',
 			'crm',
 			'search',
-			'ai',
+			'jetpack-ai',
 		);
 
 		// Add plugin action links for the core Jetpack plugin.