+++ projects/js-packages/image-guide/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.14] - 2024-11-28
+### Changed
+- Updated package dependencies. [#40060]
+
 ## [0.5.13] - 2024-11-04
 ### Added
 - Enable test coverage. [#39961]
@@ -132,6 +136,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Minor package.json change - removing private entry.
 
+[0.5.14]: https://github.com/Automattic/jetpack-image-guide/compare/v0.5.13...v0.5.14
 [0.5.13]: https://github.com/Automattic/jetpack-image-guide/compare/v0.5.12...v0.5.13
 [0.5.12]: https://github.com/Automattic/jetpack-image-guide/compare/v0.5.11...v0.5.12
 [0.5.11]: https://github.com/Automattic/jetpack-image-guide/compare/v0.5.10...v0.5.11